fix for https://issues.apache.org/activemq/browse/AMQ-2291 - jdbc sequence changed to long
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@784862 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/store/jdbc/Statements.java
@@ -30,7 +30,7 @@ public class Statements {
     protected String binaryDataType = "BLOB";
     protected String containerNameDataType = "VARCHAR(250)";
     protected String msgIdDataType = "VARCHAR(250)";
-    protected String sequenceDataType = "INTEGER";
+    protected String sequenceDataType = "BIGINT";
     protected String longDataType = "BIGINT";
     protected String stringIdDataType = "VARCHAR(250)";
     protected boolean useExternalMessageReferences;
